Attempt resolve conflict
