+++ app/src/main/java/com/example/gymsmart/components/Workouts.kt
@@ -1,5 +1,4 @@
-package com.example.anew
-
+package com.example.gymsmart
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
